requirements.txt
fastapi==0.111.1
sqlalchemy
alembic
pydantic-settings
uvicorn[standard]
gunicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
requests
mammoth
pyjwt